refactor(angular): file structure
