+++ src/angular/src/public-api.ts
@@ -1,6 +1,6 @@
 /*
  * Public API Surface of angular
  */
-export * from './lib/swiper.module';
-export * from './lib/swiper/swiper.component';
-export * from './lib/swiper-slide/swiper-slide.directive';
+export * from './swiper.module';
+export * from './swiper.component';
+export * from './swiper-slide.directive';
