+++ src/angular/src/swiper-slide.directive.ts
