+++ src/angular/src/swiper.component.html
