+++ src/angular/src/swiper.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { SwiperComponent } from './swiper/swiper.component';
-import { SwiperSlideDirective } from './swiper-slide/swiper-slide.directive';
+import { SwiperComponent } from './swiper.component';
+import { SwiperSlideDirective } from './swiper-slide.directive';
 @NgModule({
   declarations: [SwiperComponent, SwiperSlideDirective],
   exports: [SwiperComponent, SwiperSlideDirective],
