+++ src/angular/src/utils/utils.ts
